backend/requirements.txt
python-multipart
fastapi
uvicorn[standard]
google-genai
python-dotenv
pydantic-settings
asyncpg
websockets
httpx
pymupdf
python-docx
reportlab
